# PoC: Batch Insert Hosts